Rename into reset
