+++ fractal_server/app/routes/admin/v2/__init__.py
@@ -14,7 +14,7 @@ from .sharing import router as sharing_router
 from .task import router as task_router
 from .task_group import router as task_group_router
 from .task_group_lifecycle import router as task_group_lifecycle_router
-from .task_group_recollect import router as task_group_recollect_router
+from .task_group_reset import router as task_group_reset_router
 from .users_csv import router as users_csv_router
 
 router_admin = APIRouter()
@@ -24,7 +24,7 @@ router_admin.include_router(job_router, prefix="/job")
 router_admin.include_router(task_router, prefix="/task")
 router_admin.include_router(task_group_router, prefix="/task-group")
 router_admin.include_router(task_group_lifecycle_router, prefix="/task-group")
-router_admin.include_router(task_group_recollect_router, prefix="/task-group")
+router_admin.include_router(task_group_reset_router, prefix="/task-group")
 router_admin.include_router(impersonate_router, prefix="/impersonate")
 router_admin.include_router(resource_router, prefix="/resource")
 router_admin.include_router(profile_router, prefix="/profile")
