+++ fractal_server/app/schemas/v2/task_group.py
@@ -33,7 +33,7 @@ class TaskGroupActivityAction(StrEnum):
     DEACTIVATE = "deactivate"
     REACTIVATE = "reactivate"
     DELETE = "delete"
-    RECOLLECT = "recollect"
+    RESET = "reset"
 
 
 class TaskGroupCreate(BaseModel):
