+++ fractal_server/tasks/v2/local/__init__.py
@@ -1,5 +1,5 @@
 from .collect import collect_local  # noqa
-from .recollect import recollect_local  # noqa
+from .reset import reset_local  # noqa
 from .collect_pixi import collect_local_pixi  # noqa
 from .deactivate import deactivate_local  # noqa
 from .deactivate_pixi import deactivate_local_pixi  # noqa
